Merge in QA/operator-v2 from feature/add-new-crossdock-movement-schedule to master

* commit '2c838fbc93be5b779d52964a69aa3fe4630e0e25':
  reverting the method to normalisedate method
  reverting the method to normalisedate method
  review comment fix
  review comment fix
  revert changes
  review commment fixes
  review comment fixes
  added slf4j logger - review comment fix
  data changes in feature
  removed unused elements declaration
  added format date fix for Shiment inbound scanning
  fix for open shipment details
  removed unused elements declaration
  fix for format dates in shipment info page
  added format date fix for Shiment inbound scanning
  added retry for add movement schedule page

+++ src/test/java/co/nvqa/operator_v2/cucumber/glue/ShipmentInboundScanningSteps.java
@@ -37,9 +37,8 @@ public class ShipmentInboundScanningSteps extends AbstractSteps {
     retryIfRuntimeExceptionOccurred(() ->
     {
       try {
-        int shipmentId = get(KEY_CREATED_SHIPMENT_ID);
+        Long shipmentId = (Long)get(KEY_CREATED_SHIPMENT_ID);
         final String finalHub = resolveValue(hub);
-
         scanningPage.inboundScanning(shipmentId, label, finalHub);
       } catch (Throwable ex) {
         NvLogger.error(ex.getMessage());
